moving registration code to a threadpool
