+++ relay/Cargo.toml
@@ -23,3 +23,6 @@ anyhow = "1.0.32"
 os_pipe = "0.9.2"
 libc = "0.2.77" # splice syscall
 daemonize = "0.4.1"
+lazy_static = "1.4.0"
+threadpool = "1.8.1"
+ipc-channel = "0.11.0"
